backend/requirements.txt




























# FastAPI and ASGI server
fastapi==0.110.0
uvicorn==0.27.1

# Database
sqlalchemy==2.0.27
alembic==1.13.1

# Data processing
pandas==2.2.0
openpyxl==3.1.2
lxml==5.1.0

# Security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-multipart==0.0.9

# Environment variables
python-dotenv==1.0.1

# Testing
pytest==7.4.4
httpx==0.27.0



























